backend/requirements.txt
Flask
Flask-Cors
python-dotenv
openai==0.28
gunicorn
prometheus-flask-exporter